ADS507/requirements.txt
mysql-connector-python>=8.0
sqlalchemy>=2.0
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
jupyterlab>=4.0
pytest>=7.0
flake8>=6.0
watchdog>=3.0
